+++ core/search/nested_filter.cpp
@@ -31,7 +31,7 @@
 #include "search/prepared_state_visitor.hpp"
 #include "search/prev_doc.hpp"
 #include "search/score.hpp"
-#include "search/sort.hpp"
+#include "search/scorer.hpp"
 #include "utils/attribute_helper.hpp"
 #include "utils/type_limits.hpp"
